예외처리 세분화 및 Swagger Examples 추가

+++ src/main/java/com/owl/payrit/domain/repaymenthistory/service/RepaymentHistoryService.java
@@ -3,20 +3,17 @@ package com.owl.payrit.domain.repaymenthistory.service;
 import com.owl.payrit.domain.member.entity.Member;
 import com.owl.payrit.domain.promissorypaper.entity.PaperStatus;
 import com.owl.payrit.domain.promissorypaper.entity.PromissoryPaper;
-import com.owl.payrit.domain.promissorypaper.exception.PromissoryPaperException;
-import com.owl.payrit.domain.promissorypaper.service.PromissoryPaperService;
-import com.owl.payrit.domain.repaymenthistory.dto.request.RepaymentCancelRequest;
 import com.owl.payrit.domain.repaymenthistory.dto.request.RepaymentRequest;
 import com.owl.payrit.domain.repaymenthistory.entity.RepaymentHistory;
+import com.owl.payrit.domain.repaymenthistory.exception.RepaymentErrorCode;
+import com.owl.payrit.domain.repaymenthistory.exception.RepaymentException;
 import com.owl.payrit.domain.repaymenthistory.repository.RepaymentHistoryRepository;
-import com.owl.payrit.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
-import java.util.Optional;
 
 @Service
 @Slf4j
@@ -47,24 +44,28 @@ public class RepaymentHistoryService {
         LocalDate repaymentEndDate = paper.getRepaymentEndDate();
 
         if(!paper.getPaperStatus().equals(PaperStatus.COMPLETE_WRITING)){
-            throw new PromissoryPaperException(ErrorCode.REPAYMENT_STATUS_ERROR);
+            throw new RepaymentException(RepaymentErrorCode.REPAYMENT_STATUS_ERROR);
         }
 
         if(!paper.getCreditor().equals(member)) {
-            throw new PromissoryPaperException(ErrorCode.REPAYMENT_ONLY_ACCESS_CREDITOR);
+            throw new RepaymentException(RepaymentErrorCode.REPAYMENT_ONLY_ACCESS_CREDITOR);
         }
 
         if(requestDate.isBefore(repaymentStartDate) || requestDate.isAfter(repaymentEndDate)) {
-            throw new PromissoryPaperException(ErrorCode.REPAYMENT_NOT_VALID_DATE);
+            throw new RepaymentException(RepaymentErrorCode.REPAYMENT_NOT_VALID_DATE);
         }
 
         if(paper.getRemainingAmount() < repaymentRequest.repaymentAmount()) {
-            throw new PromissoryPaperException(ErrorCode.REPAYMENT_AMOUNT_OVER);
+            throw new RepaymentException(RepaymentErrorCode.REPAYMENT_AMOUNT_OVER);
         }
     }
 
     @Transactional
-    public void remove(RepaymentHistory repaymentHistory) {
+    public void remove(Member member, PromissoryPaper paper, RepaymentHistory repaymentHistory) {
+
+        if(!member.equals(paper.getCreditor())) {
+            throw new RepaymentException(RepaymentErrorCode.REPAYMENT_ONLY_ACCESS_CREDITOR);
+        }
 
         repaymentHistoryRepository.delete(repaymentHistory);
     }
@@ -72,6 +73,6 @@ public class RepaymentHistoryService {
     public RepaymentHistory getById(Long historyId) {
 
         return repaymentHistoryRepository.findById(historyId).orElseThrow(
-                () -> new PromissoryPaperException(ErrorCode.REPAYMENT_HISTORY_NOT_FOUND));
+                () -> new RepaymentException(RepaymentErrorCode.REPAYMENT_HISTORY_NOT_FOUND));
     }
 }
